+++ daemon/execdriver/native/template/default_template.go
@@ -2,8 +2,8 @@ package template
 
 import (
 	"github.com/dotcloud/docker/pkg/apparmor"
-	"github.com/dotcloud/docker/pkg/cgroups"
 	"github.com/dotcloud/docker/pkg/libcontainer"
+	"github.com/dotcloud/docker/pkg/libcontainer/cgroups"
 )
 
 // New returns the docker default configuration for libcontainer
